refactor: HttpRequest의 parameters를 StringUtils로 이동
HttpRequest의 역할 분리
