+++ src/test/java/http/HttpRequestTest.java
@@ -12,9 +12,6 @@ import java.net.URI;
 import java.net.URISyntaxException;
 import java.net.http.HttpClient;
 import java.util.List;
-import java.util.Map;
-
-import static org.assertj.core.api.Assertions.assertThat;
 
 class HttpRequestTest {
 
@@ -62,49 +59,6 @@ class HttpRequestTest {
             softAssertions.assertThat(httpRequest.version()).isEqualTo(HttpClient.Version.HTTP_1_1);
             softAssertions.assertThat(httpRequest.mime()).isEqualTo(Mime.PNG);
         });
-
-    }
-
-    @Test
-    @DisplayName("URI 쿼리스트링이 맵으로 변환된다.")
-    void convertQueryToMap() throws URISyntaxException, IOException {
-        //given
-        List<String> strings = List.of("GET /user/create?userId=javajigi&password=password&name=%EB%B0%95%EC%9E%AC%EC%84%B1&email=javajigi%40slipp.net HTTP/1.1",
-                "Host: localhost:8080",
-                "Connection: keep-alive",
-                "Accept: */*");
-        BufferedReader bufferedReader = stringListToBufferedReader(strings);
-        HttpRequest httpRequest = new HttpRequest(bufferedReader);
-
-        //when
-        Map<String, String> parameters = httpRequest.parameters(httpRequest.uri().getQuery());
-
-        //then
-        SoftAssertions.assertSoftly(softAssertions -> {
-            softAssertions.assertThat(parameters.get("userId")).isEqualTo("javajigi");
-            softAssertions.assertThat(parameters.get("password")).isEqualTo("password");
-            softAssertions.assertThat(parameters.get("name")).isEqualTo("박재성");
-            softAssertions.assertThat(parameters.get("email")).isEqualTo("javajigi@slipp.net");
-        });
-
-    }
-
-    @Test
-    @DisplayName("쿼리스트링이 없으면 빈 맵을 반환한다.")
-    void returnEmptyMapIfQueryNotExist() throws URISyntaxException, IOException {
-        //given
-        List<String> strings = List.of("GET /index.html HTTP/1.1",
-                "Host: localhost:8080",
-                "Connection: keep-alive",
-                "Accept: */*");
-        BufferedReader bufferedReader = stringListToBufferedReader(strings);
-        HttpRequest httpRequest = new HttpRequest(bufferedReader);
-
-        //when
-        Map<String, String> parameters = httpRequest.parameters(httpRequest.uri().getQuery());
-
-        //then
-        assertThat(parameters).isEmpty();
     }
 
     private BufferedReader stringListToBufferedReader(List<String> strings) {
